Renamed the file to linear.regression
