+++ shared/estimators/linear.regression.ts
