+++ shared/pipeline/pipeline.test.ts
@@ -2,7 +2,7 @@ import { assert } from 'chai'
 import Pipeline from './pipeline'
 import { tensor1d } from '@tensorflow/tfjs-core'
 import { tensorEqual } from '../utils'
-import LinearRegression from '../estimators/linear.model'
+import LinearRegression from '../estimators/linear.regression'
 import SimpleImputer from '../impute/simple.imputer'
 import MinMaxScaler from '../preprocessing/scalers/min.max.scaler'
